Refactor: move crs_to_srid to tools.crs module
